correct destroy signalR

+++ Angular/src/app/shared/bia-shared/feature-templates/crud-items/views/crud-items-index/crud-items-index.component.ts
@@ -159,7 +159,10 @@ export class CrudItemsIndexComponent<CrudItem extends BaseDto> implements OnInit
     }
     else
     {
-      this.crudItemService.signalRService.destroy();
+      if (manualChange)
+      {
+        this.crudItemService.signalRService.destroy();
+      }
     }
   }
